Signed by Jacob Newland
While my roles are framework and development, I recognize that I will be working on whichever parts of this project that I'm able to and am available for after some discussion with my teammates.  Framework is the initial plan and decisions about what this will look like, its useage, and the user experience on the site.  I intend to do this with the help of LucidChart, Vertabelo, and Trello, along with any other software useful in this front and anything I write or draw.  Deployment will entail all steps require to make our site accessible to Indivisible Arkansas and their users.  I intend to do this through the release of updates and patches for Indivisible Arkansas to see throughout our development process so that they may comment on what they like/dislike and what we could change before our final release.  With the help of my teammates, we will leave our client completely satisfied with our site by our final deadline.  I will also ensure that I help any of my teammates with any issue on this project in any way I can.